Adjust Color functions for multisampling
Adjusted write_color and write_color_str to support multiple samples per pixel, normalizing to the correct range by diving by the samples made. In the process, write_color_str was made to mutate the fmt::Write passed in without taking ownership, making it more in line with write_color. To ensure the end colors are within the correct range, a clamp function was added to util.