Implement Lambertian as a Material
Implmeneted support for handling all Materials for ray color and scattering, and implemented the Lambertian as a Material type. Removed attenuation and scattered parameters from Material::scatter as those were to be used for returning through reference parameters. Instead, a tuple is returned containing the attenuation and the scattered ray.