Create ConstantMaterial volumes
Created a ConstantMaterial hittable with uses a boundary hittable and a material which is applied based on the density throughout the whole boundary, as well as an Isotropic Material, which together can be used to make fog or smoke. These were used to make a version of the Cornell Box using smoke and fog boxes instead of the standard boxes.